// models/Appointment.ts
export class Appointment {
    firm: String = "";
    startDate: Date = new Date();
    squareFootage: Number = 0;
    privatePublicGarden: Number = 0;
    poolFootage: Number = 0;
    greenFootage: Number = 0;
    sittingFootage: Number = 0;
    fountainFootage: Number = 0;
    sittingNumber: Number = 0;
    description: String = "";
    services: String = "";
    accepted: boolean = false;
    owner: String = "";
    made: Date = new Date();
    poolNumber: number = 0;
    fountainNumber: number = 0;
    look: Object[] = []
}